requirements.txt
streamlit==1.29.0
pandas==2.0.3
plotly==5.18.0
scikit-learn==1.3.0
google-generativeai==0.3.1
numpy==1.24.3